Added models for Species and CollisionModel to pdep app.
This is the first step toward enabling wizard-based creation of MEASURE
input files via a web interface. The attributes in both Species and
CollisionModel reflect this intended purpose. (Still missing from
Species are models representing the molecular degrees of freedom and
thermodynamics.)

Note that each CollsionModel corresponds to exactly one Species and vice
versa. To indicate this relationship in Django terms, we use a
OneToOneField.