Added forms for Species and CollisionModel to pdep app.
This takes advantage of Django's functionality for creating forms
semi-automatically from models. We also use the inline formset factory
to allow for editing the collision model within the species it belongs
to. Note in particular how we override the form of the
CollisionModelFormSet and that we store this formset on the SpeciesForm
so that it is available to the template for rendering.