Added models for various molecular degrees of freedom.
The models for Translation, RigidRotor, HarmonicOscillator, and
HinderedRotor largely correspond to their counterparts in RMG Py, except
being defined using Django's model framework. (The HinderedRotor model
also does not currently accept a Fourier series for the potential.) Note
the use of ForeignKeys to associate each with a States model (a
many-to-one relationship).