Added start page for choosing an input method for new MEASURE input.
This page uses the start.html template, the pdep.views.start view, and
the InputMethodForm form. It represents a simple radio select to choose
between using the wizard to create a new input file or uploading an
existing input file. On POST the view redirects you to the appropriate
URL based on your choice.