Now creating new Network object in pdep start view.
When the "Start Here" link is invoked, the pdep app will now create a
new Network object, save it in the database, and redirect you to the
main page for that network (a view which hasn't been implemented yet).
The Network object is a new Django model for representing an individual
pressure-dependent reaction network within Django.

For the moment, an integer index is used as the preferred key for the
Network. In the future we may wish to use a randomly-generated text key
(a la Doodle) instead, to ensure that individual users can't
accidentally change each other's networks.