Added first version of networkIndex view.
This view is intended to be the main page through which users interact
with individual networks. Currently this view simply provides various
options for adding/editing input parameters (none of which are
implemented yet). A title attribute was also added to the Network class
so that we have a way to refer to the network other than its preferred
key.