Added functionality for running MEASURE to the appropriate views.
This includes normal running of MEASURE to generate output and also PES
drawing. Note that some refactoring of the PES drawing functions has
been done, primarily to split the PNG, PDF, and SVG versions of each
method apart.

I'm happy to see that the Cythonized modules in MEASURE seem to work
fine with the rest of Django.