Added QuantityField custom model field to pdep app.
This custom model field provides a nice way of storing a physical
quantity: a value with associated units. It goes nicely with
NumberUnitsField and its descendants. Since the value and units of the
physical quantity are very closely tied together, this combination makes
a lot of sense. For example, using QuantityField means that you don't
need two model fields for each physical quantity, which should make the
implementation that much more straightforward.

The goal in creating QuantityField was to make the ModelForm form
creation as automatic as possible by moving any needed configuration
options into the model.