Added Species model to pdep app.
The Species model is a Django model for representing a species in a
network. It currently has four fields: the network the species belongs
to, a unique label for the species, a structure representation, and the
ground-state energy.

A custom field class EnergyField derived from NumberUnitsField was
created for use with the E0 attribute. It provides the user with several
options for energy units.