Added form, view, and template for adding species to a network.
The SpeciesForm is a modelform based on the Species model. Since we need
many species in a network, this is made into an inline formset
SpeciesFormSet. The template networkWizardSpecies.html provides a custom
layout of this formset as a table, where each row corresponds to one
species. A form button is available for adding extra rows if needed for
additional species.

This is the first step in the wizard for creating a MEASURE input file.